Motor plate refined.
I've refined the motor plate to suit the needs of the situation. It now
is the right size and can be adjusted as needed. The chassis also had
lugs for the motor plate. I made some changes to the base plate as well
as I hone in it's placement. I think it's in the right placement I just
need to now add lugs for it's screws. I'll also need to adjust it to be
out of the way of the chute. Maybe. We'll see how that piece comes
together.